Dev zz

+++ src/main/java/com/bo/ManufactureBO.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class ManufactureBO {
-    private Long manId;
+    private Integer manId;
     private String manName;
 }
